Added syncability to DirectoryEntry.getDirectory.
This includes support for creation/retrieval of a directory at a path.  Drive doesn't support paths, so recursive traversal is necessary.

Also generalized directory creation (previously, only the creation of the syncable app directory existed).
Also assigned an error callback to a method in cases when it didn't exist.
Also added tests.