Removed syncing when a file is merely retrieved.
If DirectoryEntry.getFile is called with options.create set to false, the file is being retrieved; since nothing is changing, no sync is necessary.

If options.create is set to true, the file may still not be created (if the file already exists), and a sync will still occur.  Figuring out a way to prevent this is in the cards, but this change makes things strictly better than they were.