Auto detect and install Google Play Services.
If you're on Android, this CL will have MCA auto-detect if Google Play
Services is installed (except on Windows). If it is installed it will
auto-install the serivce into the Android project. If it isnt' installed
and you're trying to do an Android build it will cause a fatal error.