Add default iOS7 preference to push statusbar down
This will leave apps with a black statusbar (just like iOS6), but will
make fixed positioning and top-aligned elements looking as expected
out-of-the-box.  Since chrome apps don't deal with this issue, we think
this is a better default setting.

It looks less attractive than ideal on iOS7, but apps can easily change
the preference in config.xml if they wrote their app to adjust.