Set both url and uri in config for cordova create
With changes related to CB-6776 cordova now expects the urls for custom
platform files to be called url rather than uri. Setting both here
so that the current released version of cordova-lib and the future ones
can both use it. Should remove the uri once we release the next version
of lib.