correction to callback invokation.
This was allegedly expected to call `callback` with any arguments beyond `msg` and `callback`. The code being corrected was calling `Array.prototype.slice(...)` which attempts to slice `Array.prototype` and not `slice`'s first argument. I think the intention was to call `Array.prototype.slice` in the context of `arguments`, passing `2` as the 1st parameter.