Allow versions with 4 dot separated numbers
Fixes #401

Any version names with 1 to 4 dot separated numbers are now allowed. All of
the following are valid: 0.0.1.0, 2, 42, 3.14, 1.2.3

For version code calculation only the first 3 numbers will be considered the
fourth is ignored. If there are less than three numbers, extra .0 or .0.0 is
appended (this is done in cordova-lib).

Also removed an old workaround for a cordova-lib issue that will be in with the
next release.